Build container for dependencies.R
Adds a github action to build a `hdr-bgnn/minnow-traits-r-env`
container based on `Scripts/dependencies.R` when this file changes
in the main branch. When this file changes on the main branch
a new container will be created with the current date/time.

The container created will follow this pattern:
```
ghcr.io/hdr-bgnn/minnow-traits-r-env:<year>-<month>-<day>-<hour>-<minute>
```

A blank line was added to the end of Scripts/dependencies.R so an
initial container will be build when this commit is merged.